halt mapping sorting

+++ DTrace.pl
@@ -510,10 +510,6 @@ if (exists($runlevel{$runlevels}) or exists($runTask{'mapping'}) or exists($runT
     }
   }
 
-  if ($options{'seqType'} =~ /WGS/) {
-    print STDERR "stop for disk space checking\n";
-    exit 0;
-  }
 
   unless (-s "$finalBam" and !(exists($runTask{'indelRealignment'}) or exists($runTask{'MarkDuplicates'}) or exists($runTask{'recalMD'}) or exists($runTask{'BaseRecalibration'})) ) {    #processing bam
     if (-s "$rawBam" and !(-s "$sortedBam")) {     #must sort
@@ -522,6 +518,12 @@ if (exists($runlevel{$runlevels}) or exists($runTask{'mapping'}) or exists($runT
       $cmd = bwaMapping->bamIndex($confs{'samtoolsBin'}, $sortedBam);     #index it
       RunCommand($cmd,$options{'noexecute'},$options{'quiet'});
     }
+
+    if ($options{'seqType'} =~ /WGS/) {
+      print STDERR "stop for disk space checking\n";
+      exit 0;
+    }
+
     if (-s "$rawBam" and -s "$sortedBam") {
       my $cmd = "rm $rawBam -f";
       RunCommand($cmd,$options{'noexecute'},$options{'quiet'});
